+++ code/system/sys.util/src/Path/Path.trim.mts
