+++ code/system/sys.util/src/Path/index.mts
@@ -1,0 +1,1 @@
+export * from './Path.mjs';
